CodeCleanup: Moved and renamed FileManager to Dashboard
